+++ src/main/java/daw2a/gestion_alimentos_api_rest/dto/ubicacion/UbicacionDTO.java
@@ -7,6 +7,6 @@ public class UbicacionDTO {
     private Long id;
     private String descripcion;
     private String tipoUbicacion;
-    private int capacidad;
+    private Long capacidad;
 
 }
